Update pause manager

+++ pkg/execution/pauses/manager.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/google/uuid"
 	"github.com/inngest/inngest/pkg/execution/state"
 	"github.com/inngest/inngest/pkg/logger"
 )
@@ -14,9 +15,14 @@ var BlockFlushQueueName = "block-flush"
 
 var defaultFlushDelay = 10 * time.Second
 
-// StateBufferer transforms a state.Manager into a state.Bufferer
-func StateBufferer(rsm state.Manager) Bufferer {
-	return &redisAdapter{rsm}
+type ManagerOpt func(m Manager)
+
+func WithFlushDelay(delay time.Duration) ManagerOpt {
+	return func(m Manager) {
+		if mgr, ok := m.(*manager); ok {
+			mgr.flushDelay = delay
+		}
+	}
 }
 
 // NewManager returns a new pause writer, writing pauses to a Valkey/Redis/MemoryDB
@@ -24,13 +30,28 @@ func StateBufferer(rsm state.Manager) Bufferer {
 //
 // Blocks are flushed from the buffer in background jobs enqueued to the given queue.
 // This prevents eg. executors and new-runs from retaining blocks in-memory.
-func NewManager(buf Bufferer, bs BlockStore, flusher BlockFlushEnqueuer) *manager {
-	return &manager{
+func NewManager(buf Bufferer, bs BlockStore, flusher BlockFlushEnqueuer, opts ...ManagerOpt) Manager {
+	mgr := &manager{
 		buf:        buf,
 		bs:         bs,
 		flusher:    flusher,
 		flushDelay: defaultFlushDelay,
 	}
+
+	for _, o := range opts {
+		o(mgr)
+	}
+
+	return mgr
+}
+
+// NewRedisOnlyManager is a manager that only uses Redis as a buffer, without block flushing.
+func NewRedisOnlyManager(rsm state.PauseManager) Manager {
+	return NewManager(
+		StateBufferer(rsm),
+		nil,
+		nil,
+	)
 }
 
 type manager struct {
@@ -38,6 +59,23 @@ type manager struct {
 	bs         BlockStore
 	flusher    BlockFlushEnqueuer
 	flushDelay time.Duration
+}
+
+// PauseTimestamp returns the created at timestamp for a pause.
+func (m manager) PauseTimestamp(ctx context.Context, index Index, pause state.Pause) (time.Time, error) {
+	return m.buf.PauseTimestamp(ctx, index, pause)
+}
+
+func (m manager) PauseByInvokeCorrelationID(ctx context.Context, workspaceID uuid.UUID, correlationID string) (*state.Pause, error) {
+	return m.buf.PauseByInvokeCorrelationID(ctx, workspaceID, correlationID)
+}
+
+func (m manager) WriteFlushWatermark(ctx context.Context, index Index, wm FlushWatermark) error {
+	return m.buf.WriteFlushWatermark(ctx, index, wm)
+}
+
+func (m manager) GetFlushWatermark(ctx context.Context, index Index) (*FlushWatermark, error) {
+	return m.buf.GetFlushWatermark(ctx, index)
 }
 
 func (m manager) ConsumePause(ctx context.Context, pause state.Pause, opts state.ConsumePauseOpts) (state.ConsumePauseResult, func() error, error) {
@@ -67,7 +105,7 @@ func (m manager) ConsumePause(ctx context.Context, pause state.Pause, opts state
 	// case when consuming, and always re-delete the pause.  that’s no big deal, but
 	// not the best.
 	//
-	// In the future, we could add two block indexes:  pending, and stored.  this is a
+	// In the future, we could add two block indexes:  pending, and stofed.  this is a
 	// pain, though, because we may die when uploading pending blocks, and that requires
 	// a bit of thought to work around, so we’ll just go with double deletes for now,
 	// assuming this won’t happen a ton.  this can be improved later.
@@ -114,15 +152,32 @@ func (m manager) Write(ctx context.Context, index Index, pauses ...*state.Pause)
 		return n, err
 	}
 
+	if m.bs == nil || SkipFlushing(index, pauses) {
+		// Don't bother flushing, as this needs to be kept in the buffer.
+		return n, nil
+	}
+
 	// If this is larger than the max buffer len, schedule a new block write.  We only
 	// enqueue this job once per index ID, using queue singletons to handle these.
-	if m.bs != nil && n >= m.bs.BlockSize() {
+	if n >= m.bs.BlockSize() {
 		if err := m.flusher.Enqueue(ctx, index); err != nil {
 			logger.StdlibLogger(ctx).Error("error attempting to flush block", "error", err)
 		}
 	}
 
 	return n, nil
+}
+
+func (m manager) PauseByID(ctx context.Context, envID, pauseID uuid.UUID) (*state.Pause, error) {
+	// First, attempt to load this pause from the buffer.  Some pauses will definitely be here:
+	//
+	// - There aren't enough to flush to blocks, or we havent flushed yet.
+	// - We always keep pauses by ID for `step.invoke` and `step.waitForSignal` for fast O(1)
+	//   lookups to resolve these quickly
+	//
+	// If the pause isn't in the buffer, we check if the [env, event] index has been flushed before,
+	// and if so we attempt to load from the buffer.
+	return nil, fmt.Errorf("not implemented")
 }
 
 // PausesSince loads pauses in the bfufer for a given index, since a given time.
@@ -134,6 +189,10 @@ func (m manager) PausesSince(ctx context.Context, index Index, since time.Time) 
 	bufIter, err := m.buf.PausesSince(ctx, index, since)
 	if err != nil {
 		return nil, err
+	}
+
+	if m.bs == nil {
+		return bufIter, nil
 	}
 
 	blocks, err := m.bs.BlocksSince(ctx, index, since)
@@ -152,7 +211,7 @@ func (m manager) PausesSince(ctx context.Context, index Index, since time.Time) 
 
 // Delete deletes a pause from from block storage or the buffer.
 func (m manager) Delete(ctx context.Context, index Index, pause state.Pause) error {
-	// XXX: Potential future optimization:  cache the last written block for an index
+	// Potential future optimization:  cache the last written block for an index
 	// in-memory so we can fast lookup here:
 	//
 	// if blockID.ts > pause.ts, skip deleting from the buffer as the pause is in a block.
@@ -163,12 +222,21 @@ func (m manager) Delete(ctx context.Context, index Index, pause state.Pause) err
 	if err != nil && !errors.Is(err, ErrNotInBuffer) {
 		return err
 	}
+
+	if m.bs == nil {
+		return nil
+	}
+
 	// Always also delegate to the flusher, just in case a block was written whilst
 	// we issued the delete request.
 	return m.bs.Delete(ctx, index, pause)
 }
 
 func (m manager) FlushIndexBlock(ctx context.Context, index Index) error {
+	if m.bs == nil {
+		return nil
+	}
+
 	// Ensure we delay writing the block.  This prevents clock skew on non-precision
 	// clocks from impacting out-of-order pauses;  we want pauses to be stored in-order
 	// and pause blocks to contain ordered pauses.
